requirements.txt
astor==0.8.1
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.0
datasets==2.19.1

paddlenlp==2.6.1
librosa
